requirements.txt
openai
google-generativeai
python-dotenv
pydantic
rich
